backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
python-dotenv
PyJWT
